Refactor library and test that TH works
This commit goes all in on Haskell.nix for building. The upside is
that we now have full builds on different architectures. The downside
is that a lot has to be built when first loading the project. I feel
that it is a worthy trade off right now.

Working builds with TH

Refactor yay!